Solve backjoon 10217
Finally I solved it! This problem takes 3 days to finish off. I think
the main difficulty of this problem is to find bugs in my code, which is
mainly caused of difference between set and priority_queue. And then,
I find difference solution using BFS that is much faster and easier than
dijkstra. Thanks.